+++ image-viewer/main.js
@@ -1,3 +1,11 @@
-import viewer from './viewer'
+import { $ } from 'bliss'
+import config from './config'
+import initViewer from './viewer'
 
-viewer()
+$.ready().then(async () => {
+  const hostConfig = await config.getHostConfig()
+
+  initViewer(hostConfig.hosts)
+
+  config.init(hostConfig)
+})