+++ pornolab-enhancer/features/config/index.js
@@ -5,35 +5,37 @@ import store from 'store'
 
 import configCSS from './styles.css'
 
-export default (function () {
+export default (function() {
   const KEYS = {
     tags: 'tags',
     similar: 'similar',
     pager: 'pager',
-    download: 'download',
-    image: 'image'
+    download: 'download'
   }
 
-  function getRow (label, storeKey, checked) {
+  function getRow(label, storeKey, checked) {
     return $.create('label', {
       className: 'config-form__label',
-      contents: [{
-        tag: 'input',
-        type: 'checkbox',
-        className: 'config-form__checkbox js-config-checkbox',
-        checked,
-        value: storeKey
-      }, label]
+      contents: [
+        {
+          tag: 'input',
+          type: 'checkbox',
+          className: 'config-form__checkbox js-config-checkbox',
+          checked,
+          value: storeKey
+        },
+        label
+      ]
     })
   }
 
-  function createConfigForm (params) {
+  function createConfigForm(params) {
     const button = {
       tag: 'input',
       type: 'button',
       value: 'Apply',
       events: {
-        'click': (e) => {
+        click: e => {
           document.location.reload()
           Menu.hide(e)
         }
@@ -48,7 +50,16 @@ export default (function () {
         getRow('Find similar', KEYS.similar, params[KEYS.similar]),
         getRow('Pager', KEYS.pager, params[KEYS.pager]),
         getRow('Download', KEYS.download, params[KEYS.download]),
-        getRow('Image view', KEYS.image, params[KEYS.image]),
+        {
+          tag: 'div',
+          className: 'config-form__label',
+          contents: {
+            tag: 'a',
+            target: '_blank',
+            href: 'https://github.com/shikiyoku/user-scripts#image-viewer',
+            contents: 'Try Image Viewer'
+          }
+        },
         {
           tag: 'div',
           className: 'config-form__footer',
@@ -56,18 +67,18 @@ export default (function () {
         }
       ],
       delegate: {
-        'change': {
-          '.js-config-checkbox': (e) => store.set(e.target.value, e.target.checked)
-
+        change: {
+          '.js-config-checkbox': e =>
+            store.set(e.target.value, e.target.checked)
         }
       },
       events: {
-        'mousedown': (e) => e.stopPropagation()
+        mousedown: e => e.stopPropagation()
       }
     })
   }
 
-  function createMenuLink (params) {
+  function createMenuLink(params) {
     document.body.appendChild(createConfigForm(params))
 
     const container = $('#main-nav td')
@@ -83,27 +94,27 @@ export default (function () {
     // Show menu on click
     const $menuLink = jQuery(menuLink)
     $menuLink
-      .click((e) => {
+      .click(e => {
         e.preventDefault()
         Menu.clicked(jQuery(menuLink))
       })
       .hover(() => Menu.hovered($menuLink), () => Menu.unhovered($menuLink))
   }
 
-  function getParams () {
-    return Promise.all(Object.values(KEYS)
-      .map((key) => store.get(key, true)))
-      .then((values) => {
-        return Object.keys(KEYS).reduce((result, key, index) => {
-          result[key] = values[index]
-          return result
-        }, {})
-      })
+  function getParams() {
+    return Promise.all(
+      Object.values(KEYS).map(key => store.get(key, true))
+    ).then(values => {
+      return Object.keys(KEYS).reduce((result, key, index) => {
+        result[key] = values[index]
+        return result
+      }, {})
+    })
   }
 
   return {
     KEYS,
-    async init () {
+    async init() {
       const params = await getParams()
 
       $.ready().then(() => {