+++ pornolab-enhancer/main.js
@@ -4,7 +4,6 @@ import config from './features/config'
 import tags from './features/tags'
 import pagerCSS from './features/pager/styles.css'
 import download from './features/download'
-import imageView from './features/image-view'
 import findSimilar from './features/find-similar'
 
 config.init().then(params => {
@@ -20,10 +19,6 @@ config.init().then(params => {
 
   if (params[KEYS.download]) {
     download()
-  }
-
-  if (params[KEYS.image]) {
-    imageView()
   }
 
   if (params[KEYS.similar]) {