+++ image-viewer/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Image Viewer Changelog
 
-## 1.0.0 - 2018-04-14
+## 0.9.0 - 2018-04-22
 ## Added
-- Sepearatd from ple
+- Separate from ple
 - Click on backdrop closes view
+- Select enabled hosts per domain
+
+## Fixed
+- When image cannot be loaded open it in new tab (if `GM_openInTab` available)